ref(dashboard/admin): ManageCourses.jsx file deleted & Data transferred into Overview.jsx
